YEL-55 [test] 불필요한 코드 삭제
